feat(pkg): rework template for PEP 621

+++ hooks/post_gen_project.py
@@ -9,48 +9,70 @@ PROJECT_DIRECTORY = Path(".").absolute()
 def remove(filepath):
     target = PROJECT_DIRECTORY / filepath
 
+    if not target.exists():
+        return
+
     if target.is_dir():
         shutil.rmtree(target, ignore_errors=True)
     else:
         target.unlink()
 
 
-if __name__ == "__main__":
-
-    if "{{ cookiecutter.mkdocs }}" != "yes":
-        remove("docs")
-        remove("mkdocs.yml")
-
-    if "{{ cookiecutter.tool }}".lower() == "poetry":
-        to_remove = [
-            "requirements.txt",
-            "requirements-dev.txt",
-            "setup.cfg",
-            "setup.py",
-            ".coveragerc",
-        ]
-        for f in to_remove:
-            remove(f)
-    else:  # pbr
-        remove("pyproject.toml")
-        remove(".flake8")
-
+def main():
+    # Docker cleaning
     if "{{ cookiecutter.dockerfile }}" != "yes":
         remove("Dockerfile")
+        remove("requirements")
 
-    remove("dockerfiles")
+    # Remove licenses templates
     remove("licenses")
 
+    # Guard in case of template refresh using [cruft](https://github.com/cruft/cruft)
+    # TODO: use tempfile.gettempdir() to deduce on any platform
+    is_cruft_update = PROJECT_DIRECTORY.as_posix().startswith("/private")
+    if is_cruft_update:
+        return
+
     # git init
-    subprocess.call(["git", "init", "."], cwd=PROJECT_DIRECTORY)
+    subprocess.check_call(["git", "init", "."], cwd=PROJECT_DIRECTORY)
+
+    # git remote add origin
+    ssh_url = "git@github.com:{{ cookiecutter.github_username }}/{{ cookiecutter.project_name }}.git"
+    subprocess.check_call(
+        ["git", "remote", "add", "origin", ssh_url],
+        cwd=PROJECT_DIRECTORY,
+    )
+
+    https_url = "https://github.com/{{ cookiecutter.github_username }}/{{ cookiecutter.project_name }}.git"
+    notes = [
+        "- To create the github repository:",
+        f"  > gh repo create {https_url}",
+        "",
+        "- To commit and push generated template:",
+        "  > {}".format(
+            " && ".join(
+                [
+                    "cd {{ cookiecutter.project_name }}",
+                    "git add .",
+                    "git commit -m 'add(first-commit): templated by soft-cookiecutter'",
+                    "git push",
+                ]
+            )
+        ),
+        "",
+        "- To create your conda env:",
+        "  > conda create --name {{ cookiecutter.project_name}} python=3.10 --yes",
+    ]
+
+    print("\n".join(notes))
+
+
+if __name__ == "__main__":
+    try:
+        main()
+    except subprocess.CalledProcessError as err:
+        print(f"Failed to run '{err.cmd}' commands. ({err})")
+    except OSError:
+        print("Missing 'git' in PATH.")
 
     print(f"{PROJECT_DIRECTORY.as_posix()} created.")
-
-    if "{{ cookiecutter.tool }}".lower() == "poetry":
-        print("Notes:")
-        print("    - Make sure you got poetry ^1.2")
-        print("      > poetry --version")
-        print("    - Don't forget to install poetry version plugin")
-        print("      > poetry plugin add poetry-version-plugin")
-        print("    - Then install")
-        print("      > poetry install")
